Separate out aggregated init subcommand
